Move folders into more consistent structure
