+++ frontend/src/Components/Stats/StatsPage.tsx
@@ -1,0 +1,3 @@
+export function StatsPage(){
+    return <div>test</div>
+}
